pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vergil-sdk"
version = "0.1.0"
description = "Python SDK for the Vergil station local API"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "httpx>=0.27,<1",
    "websockets>=13,<15",
    "pyjwt>=2,<3",
    "platformdirs>=4,<5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
]

[project.scripts]
vergil = "vergil_sdk.__main__:main"
